@@ -2,10 +2,13 @@ from __future__ import annotations
 
 import asyncio
 from contextlib import suppress
+from typing import TYPE_CHECKING
 from typing import Any
 
 from pysignalr.client import SignalRClient
-from pysignalr.messages import CompletionMessage
+
+if TYPE_CHECKING:
+    from pysignalr.messages import CompletionMessage
 
 
 async def on_open() -> None:
@@ -26,14 +29,14 @@ async def on_error(message: CompletionMessage) -> None:
 
 def token_factory() -> str:
     # Replace with logic to fetch or generate the token
-    return "your_access_token_here"
+    return 'your_access_token_here'
 
 
 async def main() -> None:
     client = SignalRClient(
         url='https://api.tzkt.io/v1/ws',
         access_token_factory=token_factory,
-        headers={"mycustomheader": "mycustomheadervalue"},
+        headers={'mycustomheader': 'mycustomheadervalue'},
     )
 
     client.on_open(on_open)
