@@ -1,19 +1,74 @@
 import asyncio
+import atexit
 import logging
 import time
 from contextlib import suppress
-from unittest.async_case import IsolatedAsyncioTestCase
+from pathlib import Path
+from typing import Any
+from typing import cast
 
+import _pytest.outcomes
 import pytest
 import requests
+from docker.client import DockerClient  # type: ignore[import-untyped]
 
 from pysignalr.client import SignalRClient
-from pysignalr.exceptions import AuthorizationError, ServerError
+from pysignalr.exceptions import AuthorizationError
+from pysignalr.exceptions import ServerError
 
-logging.basicConfig(level=logging.INFO, format='%(asctime)s - %(levelname)s - %(message)s')
+logging.basicConfig(level=logging.DEBUG, format='%(asctime)s - %(levelname)s - %(message)s')
 
 
-def wait_for_server(url, timeout=30):
+def get_docker_client() -> 'DockerClient':
+    """Get Docker client instance if socket is available; skip test otherwise."""
+
+    docker_socks = (
+        Path('/var/run/docker.sock'),
+        Path.home() / 'Library' / 'Containers' / 'com.docker.docker' / 'Data' / 'vms' / '0' / 'docker.sock',
+        Path.home() / 'Library' / 'Containers' / 'com.docker.docker' / 'Data' / 'docker.sock',
+    )
+    for path in docker_socks:
+        if path.exists():
+            return DockerClient(base_url=f'unix://{path}')
+
+    raise _pytest.outcomes.Skipped(  # pragma: no cover
+        'Docker socket not found',
+        allow_module_level=True,
+    )
+
+
+@pytest.fixture(scope='module')
+async def aspnet_server() -> str:
+    """Run dummy ASPNet server container (destroyed on exit) and return its IP."""
+    docker = get_docker_client()
+
+    logging.info('Building ASPNet server image (this may take a while)')
+    docker.images.build(
+        path=Path(__file__).parent.parent.parent.joinpath('AspNetAuthExample').as_posix(),
+        tag='aspnet_server',
+    )
+
+    logging.info('Starting ASPNet server container')
+    container = docker.containers.run(
+        image='aspnet_server',
+        environment={
+            'ASPNETCORE_ENVIRONMENT': 'Development',
+            'ASPNETCORE_URLS': 'http://+:80',
+        },
+        detach=True,
+        remove=True,
+    )
+    atexit.register(container.stop)
+    container.reload()
+    ip = cast(str, container.attrs['NetworkSettings']['IPAddress'])
+
+    logging.info('Waiting for server to start')
+    wait_for_server(f'http://{ip}/api/auth/login')
+
+    return ip
+
+
+def wait_for_server(url: str, timeout: int = 30) -> None:
     """
     Waits for the server to be ready.
 
@@ -35,21 +90,12 @@ def wait_for_server(url, timeout=30):
         time.sleep(2)
 
 
-class TestPysignalr(IsolatedAsyncioTestCase):
-    @classmethod
-    def setUpClass(cls):
-        """
-        Sets up the test class by waiting for the server to be ready.
-        """
-        logging.info('Waiting for ASP.NET server to be ready...')
-        time.sleep(5)  # Small additional delay
-        wait_for_server('http://aspnet-server/api/auth/login', timeout=30)
-
-    async def test_connection(self) -> None:
+class TestPysignalr:
+    async def test_connection(self, aspnet_server: str) -> None:
         """
         Tests connection to the SignalR server.
         """
-        url = 'http://aspnet-server/weatherHub'
+        url = f'http://{aspnet_server}/weatherHub'
         logging.info('Testing connection to %s', url)
         client = SignalRClient(url)
 
@@ -64,23 +110,23 @@ class TestPysignalr(IsolatedAsyncioTestCase):
         with suppress(asyncio.CancelledError):
             await task
 
-    async def test_connection_with_token(self) -> None:
+    async def test_connection_with_token(self, aspnet_server: str) -> None:
         """
         Tests connection to the SignalR server with a valid token.
         """
-        login_url = 'http://aspnet-server/api/auth/login'
+        login_url = f'http://{aspnet_server}/api/auth/login'
         logging.info('Attempting to log in at %s', login_url)
         login_data = {'username': 'test', 'password': 'password'}
         response = requests.post(login_url, json=login_data, timeout=10)
         token = response.json().get('token')
         if not token:
-            self.fail('Failed to obtain token from login response')
+            pytest.fail('Failed to obtain token from login response')
 
-        url = 'http://aspnet-server/weatherHub'
+        url = f'http://{aspnet_server}/weatherHub'
         logging.info('Testing connection with token to %s', url)
 
-        def token_factory():
-            return token
+        def token_factory() -> str:
+            return cast(str, token)
 
         client = SignalRClient(
             url=url,
@@ -100,17 +146,17 @@ class TestPysignalr(IsolatedAsyncioTestCase):
             await task
 
         # Verify the token in the connection headers
-        self.assertIn('Authorization', client._transport._headers)
-        self.assertEqual(client._transport._headers['Authorization'], f'Bearer {token}')
+        assert 'Authorization' in client._transport._headers
+        assert client._transport._headers['Authorization'] == f'Bearer {token}'
 
-    async def test_invalid_token(self) -> None:
+    async def test_invalid_token(self, aspnet_server: str) -> None:
         """
         Tests connection to the SignalR server with an invalid token.
         """
-        url = 'http://aspnet-server/weatherHub'
+        url = f'http://{aspnet_server}/weatherHub'
         logging.info('Testing connection with invalid token to %s', url)
 
-        def invalid_token_factory():
+        def invalid_token_factory() -> str:
             return 'invalid_token'  # Simulate an invalid token
 
         client = SignalRClient(
@@ -135,14 +181,13 @@ class TestPysignalr(IsolatedAsyncioTestCase):
                 pass
 
         # Verify if the AuthorizationError was raised correctly
-        self.assertTrue(task.cancelled())
+        assert task.cancelled() is True
 
-    @pytest.mark.asyncio
-    async def test_send_and_receive_message(self) -> None:
+    async def test_send_and_receive_message(self, aspnet_server: str) -> None:
         """
         Tests sending and receiving a message with the SignalR server.
         """
-        login_url = 'http://aspnet-server/api/auth/login'
+        login_url = f'http://{aspnet_server}/api/auth/login'
         logging.info('Attempting to log in at %s', login_url)
         login_data = {'username': 'test', 'password': 'password'}
         response = requests.post(login_url, json=login_data, timeout=10)
@@ -152,11 +197,11 @@ class TestPysignalr(IsolatedAsyncioTestCase):
             raise AssertionError('Failed to obtain token from login response')
         logging.info('Obtained token: %s', token)
 
-        url = 'http://aspnet-server/weatherHub'
+        url = f'http://{aspnet_server}/weatherHub'
         logging.info('Testing send and receive message with token to %s', url)
 
-        def token_factory():
-            return token
+        def token_factory() -> str:
+            return cast(str, token)
 
         client = SignalRClient(
             url=url,
@@ -166,7 +211,7 @@ class TestPysignalr(IsolatedAsyncioTestCase):
 
         received_messages = []
 
-        async def on_message_received(arguments):
+        async def on_message_received(arguments: Any) -> None:
             user, message = arguments
             logging.info('Message received from %s: %s', user, message)
             received_messages.append((user, message))
@@ -179,7 +224,7 @@ class TestPysignalr(IsolatedAsyncioTestCase):
 
         async def _on_open() -> None:
             logging.info('Connection with token opened, sending message')
-            await client.send('SendMessage', ['testuser', 'Hello, World!'])
+            await client.send('SendMessage', ['testuser', 'Hello, World!'])  # type: ignore[list-item]
 
         client.on_open(_on_open)
 
